requirements.txt
aniso8601==9.0.1
anyio==3.6.1
astroid==2.8.4
attrs==21.2.0
black==22.1.0
blue==0.9.0
click==8.1.3
colorama==0.4.4
colorlog==6.6.0
fastapi==0.78.0
flake8==4.0.1
h11==0.13.0
httptools==0.4.0
idna==3.3
isort==5.9.3
itsdangerous==1.1.0
Jinja2==2.11.3
jsonschema==4.1.2
lazy-object-proxy==1.6.0
MarkupSafe==2.0.1
mccabe==0.6.1
mypy==0.910
mypy-extensions==0.4.3
pathspec==0.9.0
platformdirs==2.4.0
pyasn1==0.4.8
pycodestyle==2.8.0
pycryptodome==3.15.0
pycryptodomex==3.11.0
pydantic==1.9.1
pyflakes==2.4.0
pylint==2.11.1
pyrsistent==0.18.0
python-dotenv==0.20.0
pytz==2021.3
PyYAML==6.0
six==1.16.0
sniffio==1.2.0
starlette==0.19.1
toml==0.10.2
tomli==2.0.1
typing-extensions==3.10.0.2
uvicorn==0.18.2
watchfiles==0.15.0
websockets==10.3
Werkzeug==0.16.1
wrapt==1.13.2
